+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.21.21](https://github.com/miyaliunian/beansflight-json-form/compare/v0.2.1...v0.21.21) (2021-12-22)
+
+
+### Bug Fixes
+
+* **multi:** tablePanel支持是否开启多选 ([7799d32](https://github.com/miyaliunian/beansflight-json-form/commit/7799d32fe994a015b1823a49d3cf1c535dad6d57))
+
+
+### Features
+
+* **editform  detailform:** 支持具名插槽,并返回表单数据 通过displayed属性控制页面展现形式 ([5e4a2b0](https://github.com/miyaliunian/beansflight-json-form/commit/5e4a2b0b8352d6f923dd315d6e05f5f47ee1ad39))
+
+
+
 ## [0.21.20](https://github.com/miyaliunian/beansflight-json-form/compare/v0.2.1...v0.21.20) (2021-12-21)